[inductor] Make {freezing,ir}.py pass follow-imports typechecking
I used a couple of type-ignore comments in ir.py because it constructs
short-lived instances of FixedLayout and GraphModuleSerializer, just to
call a single method on them that doesn't use all their members. Making
those unused members optional would make the rest of the code a lot
messier with sprinkled `assert` statements.

ghstack-source-id: eba15bf42210c490c8cb4de548b99ea53cae5a68
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113534